+++ packages/create-block/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Bug Fixes
 
 - Use the description provided to fill the `description` field in `package.json` file in ESNext template ([#19867](https://github.com/WordPress/gutenberg/pull/19867)).
+- Ensure that values provided for slug and namespace get converted to lower case ([#19867](https://github.com/WordPress/gutenberg/pull/19867)).
 
 ### Internal